requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
frozenlist==1.7.0
git-filter-repo==2.47.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
outcome==1.3.0.post0
packaging==25.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
